--- src/demo/parallel/UnitTest.java
@@ -1,36 +1,0 @@
-package demo.parallel;
-
-import org.junit.Before;
-import org.junit.Test;
-import static org.junit.Assert.*;
-
- /**
- + * Created by pridybailo.n.a on 12.09.2016.
- + */
- public class UnitTest {
-     private Complex a;
-     private Complex b;
-     private Complex c;
-
-     @Before
-     public void setUp() throws Exception {
-        a = new Complex(1,0);
-        b = new Complex(0,1);
-        c = new Complex(1,1);
-     }
-
-     @Test
-     public void notEqual() throws Exception{
-        assertEquals(" not equals ", c, a.plus(b));
-     }
-
-     @Test
-     public void notNullRe() throws Exception {
-        assertNotEquals(0.0, a.getRe());
-    }
-
-     @Test
-     public void  notNullIm() throws Exception {
-        assertNotEquals(0.0, a.getIm());
-    }
- }
